Otimização e criação de alguns metodos

+++ include/tools.hpp
@@ -3,7 +3,7 @@
 #define TOOLS_HPP
 
 double random(double min, double max);
-int randint(double min, double max);
+int randint(int min, int max);
 void draw_separator_line(int size);
 
 #endif
