+++ src/tools.cpp
@@ -12,7 +12,7 @@ double random(double min, double max){
 }
 
 // Generate a random integer number between 'min' and 'max'
-int randint(double min, double max){
+int randint(int min, int max){
     std::random_device rd; 
     std::mt19937 gen(rd());
     std::uniform_int_distribution<int> dis(min, max);
